Add files via upload
This is a smart wristband designed to be a personal guardian.
It constantly monitors the wearer's vital signs and surroundings through various sensors. All the user's activity data is sent to a mobile app, which analyzes it and provides real-time alerts to
both the user and their caretaker in case of any danger, ensuring quick action.